samlCognitoAuthenticator/authenticator/requirements.txt
Arpeggio==1.9.2
astroid==2.3.3
attrs==19.3.0
boto3==1.11.9
botocore==1.14.9
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
defusedxml==0.6.0
dill==0.3.1.1
Django==2.2.7
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
ecdsa==0.15
envs==1.3
future==0.18.2
grpcio==1.26.0
idna==2.8
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
parver==0.2.1
pkgconfig==1.5.1
protobuf==3.11.2
psycopg2==2.8.4
pycparser==2.19
pycryptodome==3.3.1
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
pysaml2==5.0.0
python-dateutil==2.8.1
python-dotenv==0.10.5
python-jose-cryptodome==1.3.2
pytz==2019.3
requests==2.22.0
s3transfer==0.3.2
semver==2.9.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
urllib3==1.25.8
warrant==0.6.1
wrapt==1.11.2
